# eh-guid-report-view-new.user.js: optimization