**parse_taglist_response()**
- Better error handling
- More efficient tag data extraction

**append_dead_tags()**
- More efficient namespace handling
- More efficient DOM updates using DocumentFragment
- Optimized tag sorting by replacing localeCompare with direct string comparison and caching namespace priorities in an object for O(1) lookup

**create_tooltip_content()**
- Improved conditional logic 

**create_vote_table()**
- Removed unused parameter

**Other**
- Added debug logging system without overhead when disabled
- Improved documentation

**Benchmark Results (100 iterations)**

| Operation | New Avg (ms) | New Min (ms) | New Max (ms) | Old Avg (ms) | Old Min (ms) | Old Max (ms) |
| --- | --- | --- | --- | --- | --- | --- |
| Parse Tags | 4.45 | 3.00 | 11.00 | 5.29 | 4.00 | 12.00 |
| Append Tags | 0.65 | 0.00 | 2.00 | 2.29 | 0.00 | 5.00 |
| Add Tooltips | 2.18 | 1.00 | 6.00 | 2.45 | 1.00 | 10.00 |

Does it matter? Not really, but its fun. For reference the old script takes ~30ms on the same gallery to run enrichTags() alone, although obviously can't compare them 1:1.